Move `LiteralNestableInput` out of `sys.shards` pkg
This `NestableInput` implementation is not shards specific.
